sql: modify arithmetic aggregate functions
This patch modifies arithmetic aggredate functions, so after this patch
they will work according to out implicit cast rules. It means that if
argument of SUM ar AVG was INTEGER, then the result will also be
INTEGER. It the argument was DOUBLE, the result will be DOUBLE. NUMBER
values cannot participate in arithmetic operations.

Needed for #4415
Part of #6105