sql: introduce SQL built-in functions to parser
This patch moves SQL built-in functions to parser, so now names of these
functions are new keywords. Also, but to this the number of arguments
for functions with a variable number of arguments will be checked during
parsing. For functions with a constant number of arguments, it is always
checked during parsing.

Part of #6105