sql: separate function flags from functions
This patch separates function flags from function definition. This
allows us not to look for a function where we need its flags, but not
the function itself. It also allows us to search for a function after
the function arguments have been properly initialized, which will help
us to check the types of those arguments.

Part of #6105