sql: remove OP_BuiltinFunction0 and OP_AggStep0
This patch moves the initialization of sql_context out of the VDBE. This
allows us to remove the opcodes OP_BuiltinFunction0 and OP_AggStep0,
which work in a rather strange way. Moreover, due to the changes these
opcodes make to the VDBEs, it is possible that the estimated size of the
VDBE could become non-constant, which could lead to various problems.

Part of #6105