sql: remove unnecessary MEM finalization
After removing the setting of the error in the patch "sql: modify
arithmetic aggregate functions", we no longer need to finalize the MEM
that has not been finalized. The logic is this: if the MEM has not been
finalized, then something has happened, and we no longer need to
finalize it. The MEM will be freed like a normal MEM with allocated
memory.

Part of #6105