github-ci: add concurrency group
According to a huge amount of pushes we should cancel all outdated
workflows. Github Actions provides concurrency as a method of cancelling
(https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#concurrency).
Be careful: this is a beta feature.
Although we don't need to use this feature in case of several pushes to
master and releases branches, cause there are no force pushes.

Close tarantool/tarantool-qa#100