sql: rework SUBSTR() function
This patch makes SUBSTR() work according to ANSI rules for SUBSTRING()
function. Also, SUBSTR() can now work correctly with large INTEGER
values. The SUBSTR() syntax has not changed.

Part of #4145

@TarantoolBot document
Title: SUBSTR() function

SUBSTR() now works according to the ANSI rules for SUBSTRING(), but with
no syntax changes. What changed:
1) if the third argument is specified and it is less than 0, an error is
thrown;
2) if the sum of the second and third arguments is less than 1, an empty
string is returned;
3) if the second argument is less than 0, and the sum of the second and
third arguments is greater than 1, then the result is equal to the
result of SUBSTR() with the second argument equal to 1 and the third
argument equal to the sum of the original second and third arguments
minus 1.